Add new kernel allocator
This should be a drop-in replacement for the previous allocator, it is
ready to support user-space allocations and such.

This PR only replaces the old allocator with the new one, if this is
merged I'll add the features to allocate user pages and protections.

This allocator was tested running vim, 3D programs, etc on Linux and
seems to be quite solid. You can use the "build.sh" in kernel/src/malloc
to produce a malloc.so which can be LD_PRELOAD'ed to validate the
correctness of the allocator